Refactor tests + add test_invert

+++ oz/invert.h
@@ -27,7 +27,7 @@
 #include <flint/fmpq_poly.h>
 #include <flint/fmpz_mod_poly.h>
 
-void _fmpq_poly_oz_invert_approx(fmpq_poly_t f_inv, const fmpq_poly_t f, const int n, const mpfr_prec_t prec);
+void _fmpq_poly_oz_invert_approx(fmpq_poly_t f_inv, const fmpq_poly_t f, const long n, const mpfr_prec_t prec);
 
 void fmpq_poly_oz_invert_approx(fmpq_poly_t rop, const fmpq_poly_t f, const long n, const mpfr_prec_t prec, const uint64_t flags);
